requirements.txt
openai
pydantic>=2
python-dotenv